make tabs work and update all code snippets
