+++ docs/develop/local_dev.mdx
@@ -8,6 +8,7 @@ import Admonition from '@theme/Admonition';
 import TabItem from "@theme/TabItem";
 import Tabs from "@theme/Tabs";
 import {Scrollycoding} from "../../src/components/code/scrollycoding";
+import {TextAndCode} from "../../src/components/code/code/text-and-code";
 
 # Local Dev
 
@@ -16,37 +17,34 @@ import {Scrollycoding} from "../../src/components/code/scrollycoding";
 
 <Tabs groupId={"running-restate"}>
     <TabItem value={"npm"} label={"npm"}>
-        <Scrollycoding>
-            ## !!steps
+        <TextAndCode>
             Install Restate Server and run it:
 
-            ```shell !
+            ```shell !result
             npm install --global @restatedev/restate-server@latest &&
             restate-server
             ```
-        </Scrollycoding>
-        <Scrollycoding>
-            ## !!steps
+        </TextAndCode>
+        <TextAndCode>
             Or use npx, without installation:
 
-            ```shell !
+            ```shell !result
             npx @restatedev/restate-server
             ```
-        </Scrollycoding>
-        <Scrollycoding>
-            ## !!steps
+        </TextAndCode>
+        <TextAndCode>
             Install the Restate CLI via:
 
-            ```shell !
+            ```shell !result
             npm install --global @restatedev/restate@latest
             ```
-        </Scrollycoding>
+        </TextAndCode>
 
     </TabItem>
     <TabItem value={"Homebrew"}  label={"Homebrew"}>
         <Scrollycoding>
 
-            ## !!steps
+        ## !!steps
 
         Install Restate Server and run it with:
