+++ docs/develop/ts/service-communication.mdx
@@ -4,7 +4,6 @@ description: "Find out how Restate services can send requests to each other."
 ---
 
 import Admonition from '@theme/Admonition';
-import {TextAndCode} from "../../../src/components/code/code/TextAndCode";
 
 # Service Communication
