+++ docs/develop/ts/workflows.mdx
@@ -3,10 +3,8 @@ sidebar_position: 7
 description: "Find out how Restate services can send requests to each other."
 ---
 
-import clsx from "clsx";
 import Admonition from '@theme/Admonition';
 import {Scrollycoding} from "../../../src/components/code/scrollycoding";
-import {TextAndCode} from "../../../src/components/code/code/TextAndCode";
 
 # Workflows
