+++ docs/operate/registration.mdx
@@ -17,15 +17,11 @@ That way Restate knows where it runs and can invoke it.
 <Tabs groupId="interface" queryString>
 <TabItem value="cli" label="CLI">
 
-<CH.Section>
-
 Register a service endpoint URI as follows:
 
 ```shell
 restate deployments register localhost:9080
 ```
-
-</CH.Section>
 
 <details className="grey-details">
     <summary>Output</summary>
@@ -66,15 +62,12 @@ Or you can use the shorthand _`restate dp add`_.
 </TabItem>
 <TabItem value="curl" label="curl">
 
-<CH.Section>
 Register a service endpoint URI with [the Restate Admin API](/references/admin-api#tag/deployment/operation/create_deployment)) as follows:
 
 ```bash
 curl localhost:9070/deployments -H 'content-type: application/json' \
     -d '{"uri": "http://localhost:9080"}'
 ```
-
-</CH.Section>
 
     <details className="grey-details">
         <summary>Output</summary>
