+++ src/components/code/code-styling.module.css
@@ -9,3 +9,15 @@
 .codeContainer:hover .copyButton {
     display: block;
 }
+
+.codetablist {
+    background: #f6f8fa;
+    color: #57606a;
+}
+
+.codetab[aria-selected="true"] {
+    background: white;
+    color: black;
+    border-radius: 0.25rem 0.25rem 0 0;
+    box-shadow: inset 1px 0 0 0 #d1d5da, inset -1px 0 0 0 #d1d5da;
+}
