+++ src/components/code/code/text-and-code.module.css
@@ -1,0 +1,18 @@
+.text-and-code {
+    display: flex;
+    gap: 1rem;
+    margin: 1rem 0;
+    position: relative;
+}
+
+.text-and-code-content {
+    flex: 1;
+    box-sizing: border-box;
+    display: block;
+}
+
+.text-and-code-sticker {
+    flex: 0 0 60%;
+    box-sizing: border-box;
+    display: block;
+}
