--- src/components/code/code/textandcode.module.css
@@ -1,44 +1,0 @@
-.container {
-  /*display: flex;*/
-  /*gap: 1rem;*/
-  /*margin-bottom: 2rem;*/
-}
-
-.stepsContainer {
-  /*flex: 1;*/
-  /*margin-top: 8rem;*/
-  /*margin-bottom: 90vh;*/
-  /*margin-left: 0.5rem;*/
-  min-width: 14rem;
-}
-
-.step {
-  border-radius: 0;
-  padding: 24px;
-  border: 0;
-  border-left: 4px solid transparent;
-  box-shadow: inset 0 1px #e3e8ee;
-  margin: 0 0 0 -0.5rem;
-}
-
-.step[data-selected="true"] {
-  border: 0;
-  border-left: 4px solid var(--ifm-color-secondary-darkest);
-  background-color: var(--docusaurus-highlighted-code-line-bg);
-}
-
-.stepTitle {
-  margin-top: 1rem;
-  font-size: 1.25rem;
-}
-
-.sticker {
-  width: 60%;
-  max-width: 36rem;
-}
-
-.stickyCode {
-  position: sticky;
-  top: 4rem;
-  overflow: auto;
-}
