+++ src/components/code/scrollycoding.module.css
@@ -6,8 +6,8 @@
 
 .stepsContainer {
   flex: 1;
-  margin-top: 8rem;
-  margin-bottom: 90vh;
+  margin-top: 0.5rem;
+  margin-bottom: 30vh;
   margin-left: 0.5rem;
   min-width: 14rem;
 }
