+++ src/components/code/ui/tabs.tsx
@@ -4,6 +4,7 @@ import * as React from "react"
 import * as TabsPrimitive from "@radix-ui/react-tabs"
 
 import {cn} from "../../../lib/utils";
+import styles from "../code-styling.module.css"
 
 const Tabs = TabsPrimitive.Root
 
@@ -13,8 +14,8 @@ const TabsList = React.forwardRef<
 >(({ className, ...props }, ref) => (
       <TabsPrimitive.List
         ref={ref}
-        className={cn(
-          "inline-flex h-8 items-center justify-start rounded-md bg-zinc-100 text-zinc-500 dark:bg-zinc-800 dark:text-zinc-400 w-full",
+        className={cn(styles.codetablist,
+          "inline-flex h-8 items-center justify-start bg-zinc-100 text-zinc-500 dark:bg-zinc-800 dark:text-zinc-400 w-full",
           className,
         )}
         {...props}
@@ -28,7 +29,7 @@ const TabsTrigger = React.forwardRef<
 >(({ className, ...props }, ref) => (
   <TabsPrimitive.Trigger
     ref={ref}
-    className={cn(
+    className={cn(styles.codetab,
       "inline-flex items-center justify-center whitespace-nowrap rounded-sm px-3 py-1.5 text-sm font-medium ring-offset-white transition-all focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-zinc-950 focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 data-[state=active]:bg-white data-[state=active]:text-zinc-950 data-[state=active]:shadow-sm dark:ring-offset-zinc-950 dark:focus-visible:ring-zinc-300 dark:data-[state=active]:bg-zinc-950 dark:data-[state=active]:text-zinc-50",
       className,
     )}
