+++ src/css/custom.css
@@ -42,18 +42,6 @@
   height: 34px !important;
 }
 
-.ch-code-inline-mark, .ch-code-multiline-mark {
-  background: var(--docusaurus-highlighted-code-line-bg);
-}
-
-.ch-code-multiline-mark .ch-code-button {
-  display: none !important;
-}
-
-.ch-code-multiline-mark-border {
-  background: var(--ifm-color-secondary-darkest);
-}
-
 .alert--secondary {
   --ifm-alert-border-color: var(--ifm-color-secondary-darkest);
 }
@@ -447,6 +435,4 @@ details {
 pre {
   background: transparent;
 }
-
-
